Task-9: Add parameters to the query
fix comments
