+++ package.json
@@ -1,6 +1,6 @@
 {
   "type": "module",
-  "main": "server.js",
+  "main": "index.js",
   "dependencies": {
     "axios": "^1.7.7",
     "date-fns": "^4.1.0",
